requirements.txt
# --- Core framework ---
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt
django-autoslug
Pillow

# --- API / Documentation ---
drf-yasg==1.21.7

# --- Configuration & Environment ---
python-decouple==3.8
python-dotenv==1.0.0

# --- CORS ---
django-cors-headers==4.3.0

# --- Database ---
psycopg2-binary==2.9.9

# --- HTTP / Utilities ---
requests>=2.32,<3.0
PyYAML>=6.0,<7.0
simplejson==3.19.3

# --- Django internals (keep loose pins) ---
asgiref>=3.8,<4.0
sqlparse>=0.4,<1.0

# --- Testing ---
pytest==8.3.3
pytest-django==4.9.0

# --- Linting / Formatting ---
flake8==6.1.0
autopep8==2.0.4
isort==5.12.0

# --- Production server ---
uWSGI==2.0.23
gunicorn
setuptools